v4liveKit/backend/requirements.txt
# AIOOS Backend Requirements
# FastAPI + LiveKit + AI Services

# Web Framework
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9

# LiveKit SDK
livekit>=0.13.1
livekit-api>=0.6.0

# Database (use compatible httpx version)
supabase>=2.9.0

# Environment & Utils
python-dotenv>=1.0.1
pydantic>=2.6.1
pydantic-settings>=2.1.0
pytz>=2024.1

# Async support (compatible with supabase)
httpx>=0.24.0,<0.28.0
aiofiles>=23.2.1

# File Processing for RAG/Knowledge Base
pypdf>=4.0.0  # PDF parsing
openpyxl>=3.1.2  # Excel parsing
python-docx>=1.1.0  # Word document parsing
openai>=1.12.0  # OpenAI API for embeddings

# LiveKit Agents Framework (bara spawning agents)
# These are installed from the parent directory
livekit-agents>=0.8.0
# livekit-plugins-azure
# livekit-plugins-openai
# livekit-plugins-deepgram
# livekit-plugins-silero
